+++ prisma/migrations/20260330064456_make_goal_log_feilds_required/migration.sql
@@ -1,0 +1,12 @@
+/*
+  Warnings:
+
+  - Made the column `targetValue` on table `goal_logs` required. This step will fail if there are existing NULL values in that column.
+  - Made the column `actualValue` on table `goal_logs` required. This step will fail if there are existing NULL values in that column.
+  - Made the column `timeSpent` on table `goal_logs` required. This step will fail if there are existing NULL values in that column.
+
+*/
+-- AlterTable
+ALTER TABLE "goal_logs" ALTER COLUMN "targetValue" SET NOT NULL,
+ALTER COLUMN "actualValue" SET NOT NULL,
+ALTER COLUMN "timeSpent" SET NOT NULL;